f64 -> f32
Turns out f64 is a pretty big performance hit on wasm builds. This precision isn't necessary for any reason, so dial it down a notch.

+++ src/drawing.rs
@@ -59,7 +59,7 @@ impl Color {
     }
 }
 
-pub fn apply_gradient(colors: Vec<Color>, index: f64) -> Color {
+pub fn apply_gradient(colors: Vec<Color>, index: f32) -> Color {
     if colors.len() == 0 {
         return Color::rgb(0,0,0);
     }
@@ -67,9 +67,9 @@ pub fn apply_gradient(colors: Vec<Color>, index: f64) -> Color {
         return colors[0];
     }
 
-    let lower_index = (index * (colors.len() as f64)).floor().max(0.0).min((colors.len() - 1) as f64) as usize;
-    let upper_index = (index * (colors.len() as f64)).ceil().max(0.0).min((colors.len() - 1) as f64) as usize;
-    let weight = ((index * (colors.len() as f64)).fract() * 255.0) as u8;
+    let lower_index = (index * (colors.len() as f32)).floor().max(0.0).min((colors.len() - 1) as f32) as usize;
+    let upper_index = (index * (colors.len() as f32)).ceil().max(0.0).min((colors.len() - 1) as f32) as usize;
+    let weight = ((index * (colors.len() as f32)).fract() * 255.0) as u8;
 
 
     let final_color = Color::rgb(
